+++ tests/States/Proxy/AbstractProxyTest.php
@@ -686,7 +686,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
 
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('myCustomMethod', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame(AbstractProxyTest::class, $this->state1->getStatedClassOrigin());
         self::assertSame(array('foo', 'bar'), $this->state1->getCalledArguments());
     }
 
@@ -785,7 +785,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $object->publicMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('publicTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame('testCallFromOtherObject', $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
     }
 
@@ -829,7 +829,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $object->protectedMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('protectedTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -838,7 +838,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $object->publicMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('publicTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
     }
 
@@ -878,7 +878,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $proxy2->privateMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('privateTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -886,7 +886,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $proxy2->protectedMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('protectedTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -894,7 +894,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $proxy2->publicMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('publicTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
     }
 
@@ -933,7 +933,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $proxy->privateMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('privateTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -941,7 +941,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $proxy->protectedMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('protectedTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -949,7 +949,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $proxy->publicMethod();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('publicTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame($childClassName, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
     }
 
@@ -1207,7 +1207,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $closure();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('publicTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame(AbstractProxyTest::class, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
     }
 
@@ -1232,7 +1232,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $closure();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('privateTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame(AbstractProxyTest::class, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -1244,7 +1244,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $closure();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('protectedTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame(AbstractProxyTest::class, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
 
         //Build temp functions to test proxy behavior with different scope visibility
@@ -1256,7 +1256,7 @@ abstract class AbstractProxyTest extends \PHPUnit\Framework\TestCase
         $closure();
         self::assertTrue($this->state1->methodWasCalled());
         self::assertSame('publicTest', $this->state1->getMethodNameCalled());
-        self::assertSame('', $this->state1->getStatedClassOrigin());
+        self::assertSame(AbstractProxyTest::class, $this->state1->getStatedClassOrigin());
         self::assertSame(array(), $this->state1->getCalledArguments());
     }